Move tanglePrune to own file
